Moved function expand_sym to utils
