+++ moha/hamiltonians.py
@@ -4,9 +4,9 @@ import numpy as np
 
 from scipy.sparse import csr_matrix, diags, lil_matrix, hstack, vstack
 
-from .api import HamiltonianAPI, expand_sym
+from .api import HamiltonianAPI
 
-from .utils import get_atom_type, convert_indices
+from .utils import get_atom_type, convert_indices, expand_sym
 
 
 __all__ = [
